clases listas

+++ Tarjeta.java
@@ -1,31 +1,42 @@
-public class Tarjeta{
+/**
+ * Representa una tarjeta de crédito del banco asociada a un cliente.
+ * * @author Bravo Guadalupe, Suarez Martin
+ * @version 1.0
+ */
+public class Tarjeta {
     //Atributos
     private String tipo;
-    private Float limite;
+    private double limite;
+
     /**
      * Constructor de Tarjeta
      * @version 1.0
      * @author Bravo Guadalupe, Suarez Martin
      */
-    public Tarjeta(String tipo, Float limite){
+    public Tarjeta(String tipo, double limite){
         this.tipo = tipo;
         this.limite = limite;
     }
+
     public Tarjeta(){
         this.tipo = "";
-        this.limite = 0.0f;
+        this.limite = 0.0;
     }
+
     //Getters y setters
     public String getTipo(){
         return this.tipo;
     }
+
     public void setTipo(String tipo){
         this.tipo = tipo;
     }
-    public Float getLimite(){
+
+    public double getLimite(){
         return this.limite;
     }
-    public void setLimite(Float limite){
+
+    public void setLimite(double limite){
         this.limite = limite;
     }
 }
